web/ckanext-iauthfunctions/requirements.txt
datapunt-authorization==0.2.1
datapunt-authorization-levels==0.1.0
PyJWT>=1.4.2
